use float normal estimation, smooth normal estimation

+++ voxblox_ros/include/voxblox_ros/segmenter.h
@@ -93,8 +93,8 @@ class Segmenter {
 
   float canny_sigma_;
   int canny_kernel_size_;
-  double min_concavity_;
-  double max_dist_step_;
+  float min_concavity_;
+  float max_dist_step_;
 
   ros::Publisher edge_img_pub_;
   ros::Publisher segmentation_pub_;
